+++ src/annotations/items/AnnotationMarkerEllipse.h
@@ -35,6 +35,7 @@ public:
 
 protected:
 	void paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget) override;
+	QPainter::CompositionMode compositionMode() const override;
 };
 
 } // namespace kImageAnnotator