+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * New: Cut out vertical or horizontal slice of an image. ([#236](https://github.com/ksnip/kImageAnnotator/issues/236))
 * New: Middle-click on tab header closes tab. ([#280](https://github.com/ksnip/kImageAnnotator/issues/280))
 * New: Add button to fit image into current view. ([#281](https://github.com/ksnip/kImageAnnotator/issues/281))
+* New: Add button to fit image into current view. ([#281](https://github.com/ksnip/kImageAnnotator/issues/281))
+* New: Allow changing item opacity. ([#110](https://github.com/ksnip/kImageAnnotator/issues/110))
 * Fixed: Adding image effect does not send image change notification. ([#283](https://github.com/ksnip/kImageAnnotator/issues/283))
 
 ## Release 0.5.3